Minor oupdates and renaming resources
Removed the suffix from most of the resource names since they
are in their own resource space.

Updated justfile to include deleting resources so the cluster
doesn't have to be deleted and created with each iteration.
This is much faster than recreating the cluster.

The justfile now waits for resources to be brought up so that
the cluster is ready instead of receiving errors the first
few invocations of the web server.