Postgres using local storage
Instead of using network storage for Postgres, NFS or SMB, we use local
storage provisioned on a particular node.

Using kind, the pod storage is mapped to the node's local storage, and
the node's storage is mapped to the host.

The local storage path is retained, deleting Postgres from the node
still keeps the database files.