In progress. Updating gitea to install in homelab
Doesn't work. (There is a working version, but I'm not adding it yet,
since I want to clean up the code.)

Updating gitea to install into the cluster.

Not working due to various reasons, but mostly due to secret for
Postgres not being available.

Going to clean up and add the ability to reach the secret, but checking
this in first in case something goes amiss.