In cluster NFS server
Gave up on SMB after too many difficulties.
In particular, I couldn't get gitea to work
because the owner couldn't be set independently.

NFS works, but required a separate server to
be installed on the VM or elsewhere, and I
wanted the entire cluster to be self-sufficient
for development purposes.

Found an NFS server that works in cluster, though
there are better ones out there that I will pursue
later. For now this works.

This is configured for an ARM server currently
(as in a Macbook Pro), but can be adjusted to
an x64 machine by altering the image used for the
NFS server.

I am going to convert this to a Helm chart with
the appropriate selection of image.