Postgres now in the cluster with an NFS share
I used a slightly different tactic than with nginx.
I used storage class names to relate the PV and
PVC, with the name being derived from the storage
type being used (e.g. postgrfes-nfs).

The SMB backend is still there, but I wouldn't
recommend it as of now. It works fine with PG,
but didn't work great with other apps, such as
gitea.