Updated template and VM provisioning:
Added commands to the template provisioning.
* create to create a template using the given configuraiton
* delete to delte a template using the given configuration

The config file drives which templates are created or deleted.

Added cloud-init functionailty to VM provisioning. This is as of yet
untested, so I'm commiting this before going further. The new option
is `cloud_init_file`, which will upload the file to the server and
then configure the VM to use it when created.

`ssh_exec_interface.txt` is a file I created to upload to OpenAI so
that I wouldn't have to waste context on the actual script being
interpreted.